requirements.txt
apache-airflow==3.0.2
apache-airflow-core==3.0.2
apache-airflow-providers-standard==1.2.0
spotipy==2.25.1
pandas==2.1.4
python-dotenv==1.1.0